# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.Response:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  user.AuthRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - email
    - password
    - username
    type: object
  user.AuthResponse:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  user.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.LoginResponse:
    properties:
      email:
        type: string
      error:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
  wallet.CurrencyWalletResponse:
    properties:
      rates:
        additionalProperties:
          type: number
        type: object
    type: object
host: localhost:8080
info:
  contact:
    name: GitHub
    url: https://github.com/Sanchir01
  description: This is a sample server seller
  termsOfService: http://swagger.io/terms/
  title: "\U0001F680 Currency Wallet"
  version: "1.0"
paths:
  /exchanger:
    get:
      consumes:
      - application/json
      description: register user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.CurrencyWalletResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - refreshToken: []
      summary: Auth
      tags:
      - wallet
  /login:
    post:
      consumes:
      - application/json
      description: login user
      parameters:
      - description: auth body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Login
      tags:
      - user
  /register:
    post:
      consumes:
      - application/json
      description: register user
      parameters:
      - description: login body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Auth
      tags:
      - user
securityDefinitions:
  AccessTokenCookie:
    in: cookie
    name: accessToken
    type: apiKey
  RefreshTokenCookie:
    in: cookie
    name: refreshToken
    type: apiKey
swagger: "2.0"